Refactor: Move sync_activity into xero_workspace - Delete app sync_activity - Remove model workspace_activity - Flatten migrations
