--- apps/sync_activity/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class SyncActivityConfig(AppConfig):
-    name = 'sync_activity'
