+++ apps/xero_workspace/models.py
@@ -4,9 +4,9 @@ from django.db import models
 from django.db.models.signals import pre_delete, post_save
 from django.dispatch import receiver
 from django_q.models import Schedule
+from model_utils import Choices
 
 from apps.fyle_connect.models import FyleAuth
-from apps.sync_activity.models import Activity
 from apps.user.models import UserProfile
 from fyle_xero_integration_web_app.settings import BASE_DIR
 
@@ -117,22 +117,39 @@ class WorkspaceSchedule(models.Model):
         return str(self.id)
 
 
-class WorkspaceActivity(models.Model):
+class Activity(models.Model):
     """
-    Sync activity for Xero workspace  (Intermediate Table)
+    Activity information
     """
+    STATUS = Choices('in_progress', 'failed', 'success', 'timeout')
+    TRIGGERS = Choices('user', 'schedule', 'api')
+
     id = models.AutoField(primary_key=True, )
-    workspace = models.ForeignKey(Workspace, on_delete=models.CASCADE, help_text='FK to Workspace')
-    activity = models.ForeignKey(Activity, null=True, blank=True, related_name='activities', on_delete=models.CASCADE,
-                                 help_text='FK to Activity')
+    workspace = models.OneToOneField(Workspace, on_delete=models.CASCADE, help_text='FK to Workspace')
+    transform_sql = models.TextField(null=True, blank=True, help_text='Transform SQL')
+    sync_db_file_id = models.CharField(max_length=32, null=True, blank=True, help_text='SQLite database file Id')
+    status = models.CharField(choices=STATUS, max_length=20, default=STATUS.in_progress,
+                              help_text='Current status of the activity')
+    triggered_by = models.CharField(choices=TRIGGERS, default=TRIGGERS.user, max_length=20,
+                                    help_text='Activity triggered by')
+    request_data = models.TextField(null=True, blank=True, help_text='Request data')
+    response_data = models.TextField(null=True, blank=True, help_text='Response data')
+    error_msg = models.TextField(null=True, blank=True, help_text='Error message for user')  # Rename to display message
     created_at = models.DateTimeField(auto_now_add=True, help_text='Created at')
     updated_at = models.DateTimeField(auto_now=True, help_text='Updated at')
 
     def __str__(self):
         return str(self.id)
 
-    class Meta:
-        unique_together = ('workspace', 'activity',)
+    def update_status(self, message, status):
+        """
+        Update the Activity status
+        :param message:
+        :param status:
+        """
+        self.error_msg = message
+        self.status = status
+        self.save()
 
 
 @receiver(pre_delete, sender=WorkspaceSchedule, dispatch_uid='schedule_delete_signal')
